+++ tests/guardrails/test_pipeline.rs
@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use hub_lib::guardrails::api_control::{resolve_guards_by_name, split_guards_by_mode};
+use hub_lib::guardrails::setup::{resolve_guards_by_name, split_guards_by_mode};
 use hub_lib::guardrails::runner::execute_guards;
 use hub_lib::guardrails::providers::traceloop::TraceloopClient;
 use hub_lib::guardrails::types::*;
